Task-Management/customauth/serializers.py
from rest_framework import serializers
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from django.contrib.auth import get_user_model
from .utils import get_user_gravatar_link

User = get_user_model()


class UserSerializer(serializers.ModelSerializer):
    gravatar = serializers.SerializerMethodField()

    class Meta:
        model = User
        fields = ["id", "username", "email", "gravatar"]

    def get_gravatar(self, obj):
        return get_user_gravatar_link(obj.email)


class LoginSerializer(TokenObtainPairSerializer):

    def validate(self, attrs):
        data = super().validate(attrs)

        data["user"] = UserSerializer(self.user).data

        return data
